Add assert feature
Use assertions to document assumptions in various points in the code. The assertions added include checking for null values and range of index parameters and length of arrays carrying variables.

Parser is untouched as it is the first point of entry from user input and errors are mostly caused by users and are already handled.