Refactor and improve code quality
Cleaned up the zombie code that are commented out or not used anywhere from Alex.

Extracted the logic of parsing command details from the Parser to the specific Command classes. This reduces the long parse function and adheres to SLAP. Even though extracting the commands with switch is still long, it is still necessary and may be improved in future iterations.